% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifNA.R
\name{artifNA}
\alias{artifNA}
\title{Introduce MCAR Missing Values in a matrix}
\usage{
artifNA(x, miss.prop = 0.1)
}
\arguments{
\item{x}{a matrix, in which missing values are to be created.}

\item{miss.prop}{proportion of missing values}
}
\value{
a matrix with missing values
}
\description{
This function artificially introduces missing values in a data matrix under missing completely at random (MCAR) mechanism.
}
\examples{
 set.seed(3)
 x = matrix(rnorm(100),10,10)
 ## create 10\% missing values in x
 artifNA(x, 0.10)
}
\keyword{NA}

